+++ fastapi-keycloak/exceptions.py
