+++ fastapi-keycloak/model.py
